backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
pdfplumber
langchain
langchain-community
langchain-google-genai>=2.0.0
chromadb
python-dotenv
pandas
numpy
google-generativeai>=0.8.0
